refactored of see more button
