+++ frontend/src/components/Meme/Meme.jsx
@@ -1,15 +1,13 @@
-import React, { useState, useRef, useEffect } from "react";
+import React from "react";
 import "./Meme.css";
 import { useNavigate } from "react-router-dom";
 import { useDispatch } from "react-redux";
 import { SET_CURRENT_MEME } from "../../services/actions/currentMemeAction";
 import LikeTemplate from "../LikeTemplate/LikeTemplate";
+import { TagList } from "../TagList/TagList";
 
 function Meme({ elem, setIsNewMeme }) {
-  const [isMore, setIsMore] = useState(false);
-  const [allTagsHeight, setAllTagsHeight] = useState(0);
   const navigate = useNavigate();
-  const allTags = useRef(null);
   const dispatch = useDispatch();
   const onClick = () => {
     dispatch({ type: SET_CURRENT_MEME, payload: elem });
@@ -17,10 +15,6 @@ function Meme({ elem, setIsNewMeme }) {
     localStorage.setItem("currentMeme", JSON.stringify(elem));
     navigate(`/${elem.id}`);
   };
-
-  useEffect(() => {
-    if (allTags.current) setAllTagsHeight(allTags.current.clientHeight);
-  }, []);
 
   return (
     <li className="meme">
@@ -36,48 +30,7 @@ function Meme({ elem, setIsNewMeme }) {
         <div onClick={onClick} className="meme__image-hover"></div>
         <LikeTemplate id={elem.id} />
       </div>
-
-      {elem.tag.length > 0 && (
-        <div className="meme__tags-container">
-          <ul
-            className={`meme__tags ${isMore ? "meme__tags_more" : ""} `}
-            onClick={(e) => setIsMore(false)}
-          >
-            {elem.tag.map((tag, index) => {
-              return (
-                <li className="meme__tag" key={index}>
-                  #{tag.name}
-                </li>
-              );
-            })}
-          </ul>
-          <ul
-            ref={allTags} // невидимый полный список для сравнения высоты всех тегов с минимальной
-            className="meme__tags meme__tags_more"
-            style={{
-              opacity: 0,
-              zIndex: -2,
-            }}
-          >
-            {elem.tag.map((tag, index) => {
-              return (
-                <li className="meme__tag" key={index}>
-                  #{tag.name}
-                </li>
-              );
-            })}
-          </ul>
-
-          {!isMore && allTagsHeight > 24 && (
-            <button
-              onClick={(e) => setIsMore(true)}
-              className="meme__bth-see-more"
-            >
-              смотреть больше
-            </button>
-          )}
-        </div>
-      )}
+      <TagList elem={elem}></TagList>
     </li>
   );
 }
